feat(landingpage): allow to specify topics, products and new exclude
