+++ scripts/post.js
@@ -16,6 +16,7 @@ import {
   getLink,
   wrap,
   createTag,
+  extractTopicsAndProducts,
 } from '/scripts/common.js';
 
 import {
@@ -64,45 +65,8 @@ function handleImmediateMetadata() {
     window.blog.date = d && d.length > 0 ? formatLocalDate(d[0]) : '';
     if (window.blog.date) window.blog.rawDate = d[0];
   }
-  // store topics
-  const last = getSection();
-  let topics, topicContainer;
-  Array.from(last.children).forEach((i) => {
-    const r = /^Topics\: ?(.*)$/gmi.exec(i.innerText);
-    if (r && r.length > 0) {
-      topics = r[1].split(/\,\s*/);
-      topicContainer = i;
-    }
-  });
-  topics = topics
-    ? topics.filter((topic) => topic.length > 0)
-    : [];
-  if (topicContainer) {
-    topicContainer.remove();
-  }
-
-  // raw topics (i.e as written in the source document)
-  window.blog.topics = topics;
-
-  let products, productContainer;
-  Array.from(last.children).forEach((i) => {
-    const r = /^Products\: ?(.*)$/gmi.exec(i.innerText);
-    if (r && r.length > 0) {
-      products = r[1].split(/\,\s*/);
-      productContainer = i;
-    }
-  });
-
-  // raw products (i.e as written in the source document)
-  window.blog.products = products
-    ? products.filter((product) => product.length > 0)
-    : [];
-  if (productContainer) {
-    productContainer.remove();
-  }
-  if (last.innerText.trim() === '') {
-    last.remove(); // remove empty last div
-  }
+  
+  extractTopicsAndProducts();
 
   addMetaTags([{
     property: 'og:locale',
@@ -114,20 +78,6 @@ function handleImmediateMetadata() {
 }
 
 /**
- * Retrieves parents of specified topic from the taxonomy.
- */
-function getParentTopics(taxonomy, topics, category) {
-  let parentTopics = [];
-  topics.forEach((topic) => {
-    const parents = taxonomy.getParents(topic, category);
-    if (parents && parents.length > 0) {
-      parentTopics = parentTopics.concat(parents);
-    }
-  });
-  return parentTopics;
-}
-
-/**
  * Finds user facing topics to display, and adds both user and non user facing topics as meta tags.
  */
 async function handleAsyncMetadata() {
@@ -135,12 +85,12 @@ async function handleAsyncMetadata() {
   
   const allTopics = Array.from(new Set([
     ...window.blog.topics,
-    ...getParentTopics(taxonomy, window.blog.topics),
+    ...taxonomy.getParents(window.blog.topics),
   ]));
 
   const allProducts = Array.from(new Set([
     ...window.blog.products,
-    ...getParentTopics(taxonomy, window.blog.products, taxonomy.PRODUCTS),
+    ...taxonomy.getParents(window.blog.products, taxonomy.PRODUCTS),
   ]));
